requirements.txt
Django==5.0.3
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
drf-spectacular>=0.25.0
django-filter==24.3
django-cors-headers==4.9.0
psycopg[binary]==3.3.3
gunicorn==23.0.0
reportlab==4.2.0
celery==5.4.0
redis==5.0.7
django-celery-beat==2.6.0
django-celery-results==2.5.1
openpyxl==3.1.2
